memo 최적화
