+++ src/Components/Headers/Header.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
-import {StyleSheet, Text, TouchableOpacity, View} from 'react-native';
+import {StyleSheet, Text, View} from 'react-native';
 
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import type {StackParamsList} from '../../types/stackParamList';
-import {LinearGradient} from 'expo-linear-gradient';
 import {BackButton} from './BackButton';
+import {NextButton} from './NextButton';
+import {useCallback} from 'react';
 
 type Props = {
   navigation: NativeStackNavigationProp<StackParamsList, keyof StackParamsList>;
@@ -15,69 +16,53 @@ type Props = {
   disableNext?: boolean;
 };
 
-export function Header({
-  navigation,
-  title,
-  color = 'blue',
-  next,
-  onPressNext,
-  disableNext,
-}: Props) {
-  function goBack() {
-    navigation.pop();
-  }
-
-  function goNext() {
-    if (next) {
-      navigation.navigate(next);
+export const Header = React.memo(
+  ({
+    navigation,
+    title,
+    color = 'blue',
+    next,
+    onPressNext: processBeforeNext,
+    disableNext,
+  }: Props) => {
+    function goBack() {
+      navigation.pop();
     }
-  }
 
-  return (
-    <View style={styles.headerWrap}>
-      <BackButton color={color} onPress={goBack} />
-      <View style={styles.titleWrap}>
-        <Text
-          style={[
-            styles.title,
-            {color: color === 'blue' ? '#0000cc' : 'white'},
-          ]}>
-          {title}
-        </Text>
+    const goNext = useCallback(() => {
+      if (next) {
+        navigation.navigate(next);
+      }
+    }, [navigation, next]);
+
+    const processAndNext = useCallback(() => {
+      if (processBeforeNext) {
+        processBeforeNext();
+      }
+      goNext();
+    }, [goNext, processBeforeNext]);
+
+    return (
+      <View style={styles.headerWrap}>
+        <BackButton color={color} onPress={goBack} />
+        <View style={styles.titleWrap}>
+          <Text
+            style={[
+              styles.title,
+              {color: color === 'blue' ? '#0000cc' : 'white'},
+            ]}>
+            {title}
+          </Text>
+        </View>
+        {next && disableNext !== undefined ? (
+          <NextButton onPress={processAndNext} disable={disableNext} />
+        ) : (
+          <View style={{width: 50}} />
+        )}
       </View>
-      {next ? (
-        <TouchableOpacity
-          activeOpacity={0.7}
-          onPress={() => {
-            if (onPressNext) {
-              onPressNext();
-            }
-            goNext();
-          }}>
-          <LinearGradient
-            colors={['#FF6ECE', '#FF3DBD']}
-            style={{
-              width: 48,
-              height: 26,
-              borderRadius: 10,
-            }}>
-            <Text
-              style={{
-                fontFamily: 'Galmuri11',
-                color: 'white',
-                textAlign: 'center',
-                textAlignVertical: 'center',
-              }}>
-              다음
-            </Text>
-          </LinearGradient>
-        </TouchableOpacity>
-      ) : (
-        <View style={{width: 40}} />
-      )}
-    </View>
-  );
-}
+    );
+  },
+);
 
 const styles = StyleSheet.create({
   headerWrap: {
