+++ src/Components/Headers/NextButton.tsx
@@ -1,34 +1,58 @@
 import React from 'react';
-import {TouchableOpacity, Text} from 'react-native';
+import {TouchableOpacity, Text, View} from 'react-native';
 import {LinearGradient} from 'expo-linear-gradient';
 
-export const NextButton = () => {
-  return (
-    <TouchableOpacity
-      activeOpacity={0.7}
-      onPress={() => {
-        if (onPressNext) {
-          onPressNext();
-        }
-        goNext();
-      }}>
-      <LinearGradient
-        colors={['#FF6ECE', '#FF3DBD']}
-        style={{
-          width: 48,
-          height: 26,
-          borderRadius: 10,
-        }}>
-        <Text
-          style={{
-            fontFamily: 'Galmuri11',
-            color: 'white',
-            textAlign: 'center',
-            textAlignVertical: 'center',
-          }}>
-          다음
-        </Text>
-      </LinearGradient>
-    </TouchableOpacity>
-  );
+type Props = {
+  onPress: () => void;
+  disable: boolean;
 };
+
+export const NextButton = React.memo(
+  ({onPress: processAndNext, disable}: Props) => {
+    console.log(disable);
+    return (
+      <TouchableOpacity
+        activeOpacity={0.7}
+        onPress={processAndNext}
+        disabled={disable}>
+        {disable ? (
+          <View
+            style={{
+              width: 50,
+              height: 28,
+              borderRadius: 10,
+              justifyContent: 'center',
+              alignItems: 'center',
+              backgroundColor: '#FFCCEE',
+            }}>
+            <Text
+              style={{
+                fontFamily: 'Galmuri11',
+                color: 'white',
+              }}>
+              다음
+            </Text>
+          </View>
+        ) : (
+          <LinearGradient
+            colors={['#FF6ECE', '#FF3DBD']}
+            style={{
+              width: 50,
+              height: 28,
+              borderRadius: 10,
+              justifyContent: 'center',
+              alignItems: 'center',
+            }}>
+            <Text
+              style={{
+                fontFamily: 'Galmuri11',
+                color: 'white',
+              }}>
+              다음
+            </Text>
+          </LinearGradient>
+        )}
+      </TouchableOpacity>
+    );
+  },
+);
