+++ src/Components/paperStyles/PaperColorButton.tsx
@@ -1,0 +1,49 @@
+import React, {useMemo} from 'react';
+import {Image, StyleSheet, TouchableOpacity} from 'react-native';
+
+const checkMark = require('../../Assets/check.png');
+
+type Props = {
+  color: string;
+  paperColor: string;
+  setPaperColor: React.Dispatch<React.SetStateAction<string>>;
+};
+
+export const PaperColorButton = React.memo(
+  ({color, setPaperColor, paperColor}: Props) => {
+    const isSelectedColor = useMemo(
+      () => color === paperColor,
+      [color, paperColor],
+    );
+
+    return (
+      <TouchableOpacity
+        onPress={() => {
+          setPaperColor(color);
+        }}
+        activeOpacity={0.7}
+        style={[
+          styles.paperColorButton,
+          {
+            backgroundColor: color,
+          },
+        ]}>
+        {isSelectedColor && (
+          <Image style={styles.checkMarkImage} source={checkMark} />
+        )}
+      </TouchableOpacity>
+    );
+  },
+);
+
+const styles = StyleSheet.create({
+  paperColorButton: {
+    marginRight: 12,
+    height: 32,
+    width: 32,
+    borderRadius: 16,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  checkMarkImage: {height: 16, width: 16},
+});
