+++ src/Components/paperStyles/PaperStyle.tsx
@@ -8,11 +8,11 @@ interface Props {
   paperStyle?: 'dotted' | 'grid' | 'plain';
 }
 
-export const PaperStyle = ({lineColor, paperStyle}: Props) => {
+export const PaperStyle = React.memo(({lineColor, paperStyle}: Props) => {
   return (
     <View style={{position: 'relative'}}>
       {paperStyle === 'grid' && <Grid lineColor={lineColor} />}
       {paperStyle === 'dotted' && <Dotted lineColor={lineColor} />}
     </View>
   );
-};
+});
