+++ src/Components/paperStyles/PaperStyleButton.tsx
@@ -1,0 +1,102 @@
+import React, {useMemo} from 'react';
+import {Image, StyleSheet, Text, TouchableOpacity, View} from 'react-native';
+import {PAPER_STYLES} from '../../Constants/letter';
+
+const paperGrid = require('../../Assets/paper/paper_grid.png');
+const paperDotted = require('../../Assets/paper/paper_dotted.png');
+const paperPlain = require('../../Assets/paper/paper_plane.png');
+
+const checkMark = require('../../Assets/check.png');
+
+type Props = {
+  style: typeof PAPER_STYLES[number];
+  setPaperStyle: React.Dispatch<
+    React.SetStateAction<typeof PAPER_STYLES[number]>
+  >;
+  paperStyle: typeof PAPER_STYLES[number];
+};
+
+export const PaperStyleButton = React.memo(
+  ({style, paperStyle, setPaperStyle}: Props) => {
+    let source: any;
+    let name: string = '';
+
+    switch (style) {
+      case 'grid':
+        source = paperGrid;
+        name = '모눈노트';
+        break;
+      case 'dotted':
+        source = paperDotted;
+        name = '도트노트';
+        break;
+      case 'plain':
+        source = paperPlain;
+        name = '플레인노트';
+        break;
+    }
+
+    const isSelectedStyle = useMemo(
+      () => style === paperStyle,
+      [paperStyle, style],
+    );
+
+    return (
+      <TouchableOpacity
+        activeOpacity={0.7}
+        style={styles.paperStyleButton}
+        onPress={() => {
+          setPaperStyle(style);
+        }}>
+        <Image source={source} style={styles.paperStyleImage} />
+        {isSelectedStyle ? (
+          <View style={styles.checked}>
+            <Image source={checkMark} style={styles.checkMarkImageSmall} />
+          </View>
+        ) : (
+          <View style={styles.unchecked} />
+        )}
+        <Text style={styles.paperStyleName}>{name}</Text>
+      </TouchableOpacity>
+    );
+  },
+);
+
+const styles = StyleSheet.create({
+  paperStyleButton: {marginRight: 16, position: 'relative'},
+  paperStyleImage: {
+    height: 160,
+    width: 118,
+  },
+  checkMarkImageSmall: {height: 15, width: 15},
+  checked: {
+    backgroundColor: '#ff6ece',
+    position: 'absolute',
+    right: 5,
+    top: 5,
+    height: 22,
+    width: 22,
+    borderRadius: 11,
+    borderWidth: 2,
+    borderColor: '#fff',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  unchecked: {
+    backgroundColor: '#ffc7f0',
+    position: 'absolute',
+    right: 5,
+    top: 5,
+    height: 22,
+    width: 22,
+    borderRadius: 11,
+    borderWidth: 2,
+    borderColor: '#fff',
+  },
+  paperStyleName: {
+    textAlign: 'center',
+    fontFamily: 'Galmuri11',
+    fontSize: 14,
+    color: 'white',
+  },
+});
